CI: Use git compatible worksheet
The built-in worksheet doesn't deal with git versions etc very well.
While we could encode this information in the 'rev' box, it is too small.

So instead of hacking around the shortcommings of the default worksheet,
lets do our own.

One problem remains, is that SHEETPATH and SHEETNAME are the two
variables used to name a SHEET. Because they are used mutually
exclusive, the 'hack' of putting both variables in place works
reasonable, until it won't. Better would be for kicad to combine these
variables in one name.

Maybe we can make this worksheet the new default at some point.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>