lapod: Add SNV65LVDS based pod
This commit adds the SNV65LVDS based pod. This pod is based on one quad
opamp, but without the wide input range of the LMH732x parts, making it
considerable cheaper.

The BOM currently uses the SVN65LVCDS and TLV743P LDO, which limits the
input range from 1v6 to 5v5 logic levels, which for most applications is
perfectly fine for most applications, but alternative parts should fit
in the same footprint to enable 0v65 - 3v6 operation.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>