lapod: Add LMH7322 based pod
This commit adds the LMH7322 based pod. This pod is based on two dual
opamps and uses two of these chips.

This is the same chip the official rigol probe also uses. One of their
pods contains 8 chips (4 on each side).

As an option, only 1 LMH7322 can be used to only use a dual output, if
heat and cost are an issue.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>