fix: Replace SN65LVDS31D/SN65LVDS047W with SN65LVDS391PWR
We accidentally used the incorrect LVDS comparator. The difference is
subtle, but important.

The symbol came from the KiCad default library, the rest of the part
number from LCSC, but was simply the wrong one. 31D vs 391PWR. Subtle
but important.

Luckily, the footprint was almost the same, so no real changes where needed
here.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>