kibot: Switch to 1.6.5
There's a cascading bug which break the pipeline currently. One is the
fact that solder paste does not clear the underlying solder mask
underneath it explicitly. However clearing it causes the DRC to complain
where it says two layers are now clearing the same soldermask.

Upgrading to the unreleased kibot 1.6.5 fixes this, but introduces a new
problem with the kici render3d template, as the kibot changed its
interface there.

Upgrade to the 1.6.5 branch of kibot.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>